[#51677] arm64: Handle VM bit in arm_hcr_el2_eff, add arm_sctlr_eff
Previously HCR.VM would not be set when HCR.DC is set,
despite the ARM Reference Manual saying:
"the behavior of the PE when executing in a Non-secure mode other than
Hyp mode is consistent with HCR.VM being 1,
regardless of the actual value".

Additionally, this patch adds arm_sctlr_eff, which differs from arm_sctlr
in that it clears the M bit according to the spec,
based on the current HCR configuration.